Refactor: domain/ui 패키지로 분리
