+++ src/main/java/ladder/domain/LadderData.java
@@ -1,4 +1,6 @@
-package ladder;
+package ladder.domain;
+
+import ladder.ui.Input;
 
 public class LadderData {
