+++ src/main/java/ladder/domain/LineType.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public enum LineType {
