+++ src/main/java/ladder/ui/Input.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.ui;
 
 public interface Input {
